move each module to single purpose
