+++ src/PIRServer.js
@@ -1,5 +1,6 @@
+/* Module to provide HTTP/AJAX comms with a VentOS server */
 import $ from 'jquery';
-import {add_samples} from './PIRUtils';
+import {concatSamples, pircsParameter, pircsValue, pircsInterpretation} from './PIRUtils';
 import {plot} from './PlotlyPlotter';
 import axios from "axios";
 import qs from "qs";
@@ -10,57 +11,10 @@ var MAX_SAMPLES_TO_STORE_S = 2000;
 var settings = {}
 var samples = []
 
-// https://github.com/PubInv/PIRCS-pubinv-respiration-control-standard/blob/master/PIRCS.md
-const name_to_PIRCS_map = {
-  Mode: 'M',
-  Pmax: 'P',
-  Pi: 'P',
-  PEEP: 'E',
-  TV: 'V',
-  Flow: 'F',
-  RR: 'B',
-  IE: 'I',
-  FiO2: 'O',
-}
-
-const pircsValue = (par, value) => {
-  if (par === 'I') { // I:E - handle a colon if we get one!
-    const vs = value.split(":");
-    if (vs.length === 1) {
-      return 10 * parseInt(vs[0])
-    } else if (vs.length === 2) {
-      return 10 * parseInt(vs[1]) / parseInt(vs[0]);
-    }
-    else {
-      alert("INTERNAL GUI ERROR: I:E value just have colon");
-      throw new Error('Bad PIRCS IE')
-    }
-  } else if ('PBO'.includes(par)) { // pressure, RR, FiO2
-    return 10 * parseInt(value)
-  } else {
-    return parseInt(value)
-  }
-}
-
-// interpretations are: m: Minimum M: Maximum T: Target
-const pircsInterpretations = (name) => {
-  if (name === 'Pmax') {
-    return 'M'
-  }
-  else {
-    return 'T'
-  }
-
-}
-
-const PIRCS = (name, value) => {
-
-  const par = name_to_PIRCS_map[name]
-  if (!par) {
-    throw new Error('Bad PIRCS parameter name')
-  }
+const sendPIRCS = (name, value) => {
+  const par = pircsParameter(name)
   const val = pircsValue(par, value)
-  const int = pircsInterpretations(name)
+  const int = pircsInterpretation(name)
   var url = process.env.REACT_APP_DSERVER_URL + "/control/";
 
   // name = TV
@@ -97,16 +51,17 @@ export const server = {
     }
     settings = session_settings || settings
     intervalId = setInterval(
-      function() {getPIRDSData()},
+      function() {getPIRDS()},
       DATA_RETRIEVAL_PERIOD);
     console.log('starting server', intervalId)
     settings.live = true;
   },
   isLive: () => settings.live,
   settings: () => settings,
+  sendPIRCS,
 }
 
-function getPIRDSData() {
+function getPIRDS() {
   const DSERVER_URL = settings.dserverurl; //"https://ventos.dev/ventos";
   const url = DSERVER_URL + "/"+ MAX_SAMPLES_TO_STORE_S;
   // fixme change to axios
@@ -114,7 +69,7 @@ function getPIRDSData() {
     url: url,
     success: function(new_samples){
       if (new_samples) {
-        samples = add_samples(samples, new_samples, MAX_SAMPLES_TO_STORE_S);
+        samples = concatSamples(samples, new_samples, MAX_SAMPLES_TO_STORE_S);
         plot(samples);
       } else {
         alert('no samples');
