+++ src/PIRUtils.js
@@ -1,3 +1,6 @@
+/**
+ * Module to provide generic functions to support PIRCS and PIRDS data
+ */
 
 // WARNING: This is a hack...if the timestamp is negative,
 // we treat it as a limited (beyond range of sensor) measurement.
@@ -14,7 +17,7 @@ function sanitize_samples(samples) {
                       if ("string" === (typeof s.num))
                         s.num = parseInt(s.num);
                       if (s.ms < 0) {
-                        s.ms = -s.ms;
+                        s.ms = -s.ms; // fixme! what is going on here???
                       } else if (s.event === "E") {
                       }
                     }
@@ -23,25 +26,71 @@ function sanitize_samples(samples) {
   return samples;
 }
 
-export function add_samples(samples, cur_sam, MAX_SAMPLES_TO_STORE_S) {
-  cur_sam = sanitize_samples(cur_sam);
+// note: not guaranteeed to get samples in order
+export function concatSamples(samples, new_samples, MAX_SAMPLES_TO_STORE_S) {
+  new_samples = sanitize_samples(new_samples);
   var discard = Math.max(0,
-                         samples.length + cur_sam.length - MAX_SAMPLES_TO_STORE_S);
-  samples = samples.slice(discard);
-  samples = samples.concat(cur_sam);
+    samples.length + new_samples.length - MAX_SAMPLES_TO_STORE_S);
+  samples = samples.slice(discard); // fixme this should happen after the merge
+  samples = samples.concat(new_samples);
+  // fixme should be sorting new samples not old samples
   samples.sort((a,b) => a.ms < b.ms);
-  // We are not guaranteeed to get samples in order
-  // we sort them....
-  // We also need to de-dup them.
-  // This would be more efficient if done after sorting..
-
-  // I think this is de-dupeing code...
+  // de-dupe - fixme de-duping should scan lists based on timestamps
   samples = samples.filter((s, index, self) =>
     self.findIndex(t => t.ms === s.ms
-                   && t.type === s.type
-                   && t.loc === s.loc
-                   && t.num === s.num
-                   && t.event === s.event
-                   && t.val === s.val) === index);
+      && t.type === s.type
+      && t.loc === s.loc
+      && t.num === s.num
+      && t.event === s.event
+      && t.val === s.val) === index);
   return samples
 }
+
+// https://github.com/PubInv/PIRCS-pubinv-respiration-control-standard/blob/master/PIRCS.md
+export function pircsParameter(name) {
+  const name_to_PIRCS_map = {
+    Mode: 'M',
+    Pmax: 'P',
+    Pi: 'P',
+    PEEP: 'E',
+    TV: 'V',
+    Flow: 'F',
+    RR: 'B',
+    IE: 'I',
+    FiO2: 'O',
+  }
+  const par =  name_to_PIRCS_map[name]
+  if (!par) {
+    throw new Error('Bad PIRCS parameter name')
+  }
+  return par
+}
+
+export function pircsValue(par, value) {
+  if (par === 'I') { // I:E - handle a colon if we get one!
+    const vs = value.split(":");
+    if (vs.length === 1) {
+      return 10 * parseInt(vs[0])
+    } else if (vs.length === 2) {
+      return 10 * parseInt(vs[1]) / parseInt(vs[0]);
+    }
+    else {
+      throw new Error('Bad PIRCS IE')
+    }
+  } else if ('PBO'.includes(par)) { // pressure, RR, FiO2
+    return 10 * parseInt(value)
+  } else {
+    return parseInt(value)
+  }
+}
+
+// interpretations are: m: Minimum M: Maximum T: Target
+export function pircsInterpretation(name) {
+  if (name === 'Pmax') {
+    return 'M'
+  }
+  else {
+    return 'T'
+  }
+}
+
